requirements.txt
Django==1.11.2
MySQL-python==1.2.5
django-extensions-1.7.9
Pillow-4.1.1
